@@ -1,4 +1,4 @@
-class Answer < ActiveRecord::Base
+class Answer < ApplicationRecord
   belongs_to :question
 
   validates :title, :body, presence: true
