@@ -1,10 +1,10 @@
-class CreateQuestions < ActiveRecord::Migration
+class CreateQuestions < ActiveRecord::Migration[5.0]
   def change
     create_table :questions do |t|
       t.string :title
       t.string :body
 
-      t.timestamps null: false
+      t.timestamps
     end
 
     add_index :questions, :title
