@@ -1,7 +1,6 @@
 FactoryGirl.define do
   factory :answer do
     title "MyString"
-body "MyString"
+    body "MyString"
   end
-
 end
