@@ -1,7 +1,8 @@
 require 'rails_helper'
 
 RSpec.describe Answer, type: :model do
+  it { should belong_to(:question) }
+
   it { should validate_presence_of :title }
   it { should validate_presence_of :body }
-  it { should belong_to(:question) }
 end
